Misc (re #1068): fix compile errors/warnings on Symbian S60 5th ed.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@3255 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib/src/pj/errno.c
@@ -217,7 +217,7 @@ static void pj_perror_imp(int log_level, const char *sender,
 
     /* Build the title */
     len = pj_ansi_vsnprintf(titlebuf, sizeof(titlebuf), title_fmt, marker);
-    if (len < 0 || len >= sizeof(titlebuf))
+    if (len < 0 || len >= (int)sizeof(titlebuf))
 	pj_ansi_strcpy(titlebuf, "Error");
 
     /* Get the error */
